requirements.txt
eth-brownie
pandas